+++ index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var estream = require('event-stream'),
+	PluginError = require('gulp-util').PluginError,
 	eslint = require('eslint').linter,
 	util = require('./util');
 
@@ -14,16 +15,20 @@ function gulpEslint(options) {
 		var config = configHelper.getConfig(filePath);
 
 		// eslint.cli result structure (+config)
+		// remove shebangs (https://github.com/eslint/eslint/commit/109a57b2ca60bf7c195f935cddda9a90311860f3)
 		return {
 			config: config,
 			filePath: filePath,
-			messages: eslint.verify(contents, config, false)
+			messages: eslint.verify(contents.replace(/^#![^\r\n]+[\r\n]/, ""), config, false)
 		};
 	}
 
 	return estream.map(function (file, output) {
 
-		if (file.isStream()) {
+		if (util.checkForExclusion(file, configHelper)) {
+			output(null, file);
+
+		} else if (file.isStream()) {
 			// eslint is synchronous, so wait for the complete contents
 			// replace content stream with new readable content stream
 			file.contents = file.contents.pipe(util.wait(function (contents) {
@@ -32,16 +37,46 @@ function gulpEslint(options) {
 			}));
 
 		} else {
-
-			if (file.isBuffer()) {
-				file.eslint = verify(file.path, String(file.contents));
-			}
+			file.eslint = verify(file.path, file.contents.toString('utf8'));
 			output(null, file);
 
 		}
 	});
 
 }
+
+/**
+ * Fail when an eslint error is found in eslint results.
+ */
+gulpEslint.failOnError = function () {
+
+	return estream.map(function (file, output) {
+		var messages = file.eslint && file.eslint.messages || [],
+			config = file.eslint && file.eslint.config || {},
+			error = null;
+
+		messages.some(function (message) {
+			var level = message.fatal ? 2: config.rules[message.ruleId];
+			if (Array.isArray(level)) {
+				level = level[0];
+			}
+			if (level > 1) {
+				error = new PluginError(
+					'gulp-eslint',
+					{
+						name:'ESLintError',
+						fileName:file.path,
+						message: message.message,
+						lineNumber: message.line
+					}
+				);
+				return true;
+			}
+		});
+
+		return output(error, file);
+	});
+};
 
 /**
  * Wait until all files have been linted and format all results at once.
@@ -77,7 +112,7 @@ gulpEslint.formatEach = function (formatter, writable) {
 			try {
 				util.writeResults([file.eslint], formatter, writable);
 			} catch (err) {
-				error = err;
+				error = new PluginError('gulp-eslint', err);
 			}
 		}
 		output(error, file);