+++ test/fixtures/ignored.js
@@ -1,0 +1,5 @@
+(function () {
+	'use strict';
+	ignore = abc;
+	}// intentional syntax error
+});