+++ CHANGELOG.md
@@ -3,3 +3,21 @@
 ## 0.1.0
 
 * initial plugin
+
+## 0.1.1
+
+* Update dependency versions
+* Loosen version peer dependency on Gulp
+
+## 0.1.2
+
+* Update eslint version
+
+## 0.1.3
+
+* Change default formatter to 'stylish'
+* Add support for .eslintignore file
+* Skip non-JS files to play well with multi-filetype streams
+* Add "failOnError" method to stop streams when an eslint error has occurred
+* Use gulp-util's PluginError
+* Ignore shebangs in JS files